requirements.txt
Django==1.11.13
djangorestframework==3.8.2
funcsigs==1.0.2
mock==2.0.0
pbr==4.0.2
pytz==2018.4
requests==2.13.0
six==1.11.0
Vocabulary==1.0.4
yapf==0.21.0
